Add files via upload
This is essentially a default hotkey setup except with a few changes to help manage conflicts that grid menu has with the defaults. The changes are listed at the top of the file. Still to be solved is using fight command to rally from a lab - currently works with airlabs because there is no unit on "f".