fix `WASD` default keys custom camera preset
The previous version was improperly trying to unbind `WASD`+modifiers
using the `Any` modifier, which does not actually match the other
modifiers when it comes to the bindings themselves (it's merely a
runtime check). This meant previous keybinds on `Ctrl`+`WASD` /
`Shift`+`WASD` / `Alt`+`WASD` were actually still registered.

We fix this by properly removing any and all actions bound to `WASD` and
`WASD`+modifiers, and then rebinding the actions as was proposed in the
original file. As a nice bonus, with `Ctrl`+`WASD` and `Shift`+`WASD`
now properly unbound, this also allows them to use the `moveslow` and
`movefast` actions from the default keys.

We also fix the `Self-Destruct` on `Ctrl`+`B` by properly unbinding its
previous action from the default keys (`Select idle builder`), as
otherwise `Ctrl`+`B` would still select an idle builder when no unit was
selected and then `Self-Destruct` when a unit was selected, leading to
potential bad situations if someone were to double tap `Ctrl`+`B`,
especially at the beginning of a game... :DDD