rework `FAQ` and `README` files for camera presets
We refactor the FAQ entry about crafting a custom camera preset to
instead point to the camera presets `README` file, and refactor that
`README` with more details on all the existing presets provided.

Additionally, we initialize a redirection list in `custom_presets`'s
`README` in case someone manually browses to that directory, which
hopefully might grow as new types of custom presets are added.